LinuxでWarningになっている箇所の修正。

+++ Client/AnpanMMO/Source/AnpanMMO/Script/ScriptBinds.cpp
@@ -49,7 +49,7 @@ int SetFlag_Call(lua_State *pState)
 	ScriptExecuterBase * pExecuter = ExecuterPool::GetInstance().Get(Id);
 	if (pExecuter == NULL) { return 0; }
 
-	int Flag = luaL_checknumber(pState, -1);
+	int Flag = (int) luaL_checknumber(pState, -1);
 	pExecuter->SetFlag(Flag);
 	return 0;
 }
@@ -61,7 +61,7 @@ int GetFlag_Call(lua_State *pState)
 	ScriptExecuterBase * pExecuter = ExecuterPool::GetInstance().Get(Id);
 	if (pExecuter == NULL) { return 0; }
 
-	int Flag = luaL_checknumber(pState, -1);
+	int Flag = (int) luaL_checknumber(pState, -1);
 	bool bFlag = pExecuter->GetFlag(Flag);
 	lua_pushboolean(pState, bFlag);
